fix: reorganize util files
