+++ apps/google-docs/src/locations/Page/components/review/DocumentOutline.tsx
@@ -20,7 +20,6 @@ import {
 } from '@contentful/f36-components';
 import tokens from '@contentful/f36-tokens';
 import type {
-  EntryBlockGraph,
   EntryBlockGraphSourceRef,
   MappingReviewSuspendPayload,
   NormalizedDocumentContentBlock,
@@ -28,28 +27,25 @@ import type {
   NormalizedDocumentTable,
   NormalizedDocumentTablePart,
 } from '@types';
-import {
-  isBlockImageSourceRef,
-  isBlockSourceRef,
-  isTableImageSourceRef,
-  isTextSourceRef,
-} from '@types';
+import { isBlockImageSourceRef, isTableImageSourceRef, isTextSourceRef } from '@types';
 import { FileTextIcon } from '@contentful/f36-icons';
 import { MappingCard, type MappingCardData } from './MappingCard';
-import { getAnchorIdForSourceRef, resolveMarkerOffsets } from './mappingCardPositioning';
-import { type DocSegment, type Tab, orderDocument } from './buildDocument';
+import { getAnchorIdForSourceRef, resolveMarkerOffsets } from './utils/mappingCardPositioning';
+import { type DocSegment, orderDocument } from './utils/buildDocument';
+import {
+  buildListItemPresentations,
+  buildOverviewEntries,
+  buildUsageIndexes,
+  formatDisplayName,
+  getMappingCardKey,
+  type SourceUsage,
+  uniqueUsage,
+} from './utils/documentOutlineModel';
 
 interface DocumentOutlineProps {
   payload: MappingReviewSuspendPayload;
   showChrome?: boolean;
   onBack?: () => void;
-}
-
-interface SourceUsage {
-  entryIndex: number;
-  fieldType: string;
-  fieldId: string;
-  sourceRef: EntryBlockGraphSourceRef;
 }
 
 type AnchoredMappingCard = MappingCardData & {
@@ -62,90 +58,6 @@ type TextSegment = {
   highlighted: boolean;
   mappingKeys: string[];
 };
-
-type ListItemPresentation = {
-  marker: string;
-  nestingLevel: number;
-};
-
-const formatDisplayName = (value: string): string => {
-  const normalized = value
-    .replace(/([a-z0-9])([A-Z])/g, '$1 $2')
-    .replace(/[-_]+/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim();
-
-  if (!normalized) {
-    return value;
-  }
-
-  return normalized.replace(/\b\w/g, (char) => char.toUpperCase());
-};
-
-function buildUsageIndexes(entryBlockGraph: EntryBlockGraph): {
-  blockUsage: Record<string, SourceUsage[]>;
-  tablePartUsage: Record<string, SourceUsage[]>;
-  tableUsage: Record<string, SourceUsage[]>;
-} {
-  const blockUsage: Record<string, SourceUsage[]> = {};
-  const tablePartUsage: Record<string, SourceUsage[]> = {};
-  const tableUsage: Record<string, SourceUsage[]> = {};
-
-  entryBlockGraph.entries.forEach((mappingEntry, entryIndex) => {
-    mappingEntry.fieldMappings.forEach((fieldMapping) => {
-      fieldMapping.sourceRefs.forEach((sourceRef) => {
-        const usage: SourceUsage = {
-          entryIndex,
-          fieldId: fieldMapping.fieldId,
-          fieldType: fieldMapping.fieldType,
-          sourceRef,
-        };
-
-        if (isBlockSourceRef(sourceRef)) {
-          blockUsage[sourceRef.blockId] = [...(blockUsage[sourceRef.blockId] ?? []), usage];
-          return;
-        }
-
-        if (
-          !(
-            'tableId' in sourceRef &&
-            'rowId' in sourceRef &&
-            'cellId' in sourceRef &&
-            'partId' in sourceRef
-          )
-        ) {
-          return;
-        }
-
-        const tablePartKey = [
-          sourceRef.tableId,
-          sourceRef.rowId,
-          sourceRef.cellId,
-          sourceRef.partId,
-        ].join(':');
-        tablePartUsage[tablePartKey] = [...(tablePartUsage[tablePartKey] ?? []), usage];
-        tableUsage[sourceRef.tableId] = [...(tableUsage[sourceRef.tableId] ?? []), usage];
-      });
-    });
-  });
-
-  return { blockUsage, tablePartUsage, tableUsage };
-}
-
-const getMappingCardKey = (segmentId: string, usage: SourceUsage): string =>
-  `${segmentId}-${usage.entryIndex}-${usage.fieldId}`;
-
-function uniqueUsage<T extends SourceUsage>(usage: T[]): T[] {
-  const seen = new Set<string>();
-  return usage.filter((item) => {
-    const key = `${item.entryIndex}-${item.fieldId}-${item.fieldType}`;
-    if (seen.has(key)) {
-      return false;
-    }
-    seen.add(key);
-    return true;
-  });
-}
 
 function buildTextSegments(
   flattenedRuns: NormalizedDocumentFlattenedRun[],
@@ -261,48 +173,6 @@ function renderTextSegment(
   );
 }
 
-function buildListItemPresentations(
-  blocks: NormalizedDocumentContentBlock[]
-): Record<string, ListItemPresentation> {
-  const presentations: Record<string, ListItemPresentation> = {};
-  const orderedCounts = new Map<number, number>();
-
-  [...blocks]
-    .sort((left, right) => left.position - right.position)
-    .forEach((block) => {
-      if (block.type !== 'listItem' || !block.bullet) {
-        orderedCounts.clear();
-        return;
-      }
-
-      const nestingLevel = Math.max(0, block.bullet.nestingLevel ?? 0);
-
-      Array.from(orderedCounts.keys()).forEach((level) => {
-        if (level > nestingLevel) {
-          orderedCounts.delete(level);
-        }
-      });
-
-      if (block.bullet.ordered) {
-        const nextCount = (orderedCounts.get(nestingLevel) ?? 0) + 1;
-        orderedCounts.set(nestingLevel, nextCount);
-        presentations[block.id] = {
-          marker: `${nextCount}.`,
-          nestingLevel,
-        };
-        return;
-      }
-
-      orderedCounts.delete(nestingLevel);
-      presentations[block.id] = {
-        marker: nestingLevel > 0 ? '◦' : '•',
-        nestingLevel,
-      };
-    });
-
-  return presentations;
-}
-
 export const DocumentOutline = ({ payload, showChrome = true, onBack }: DocumentOutlineProps) => {
   const [selectedEntryIndex, setSelectedEntryIndex] = useState<number | null>(null);
   const [hoveredMappingKeys, setHoveredMappingKeys] = useState<string[]>([]);
@@ -342,22 +212,7 @@ export const DocumentOutline = ({ payload, showChrome = true, onBack }: Document
   );
 
   const overviewEntries = useMemo(
-    () =>
-      payload.entryBlockGraph.entries.map((graphEntry, entryIndex) => {
-        const contentTypeName = payload.contentTypes.find(
-          (contentType) => contentType.sys.id === graphEntry.contentTypeId
-        )?.name;
-        const displayName = formatDisplayName(contentTypeName ?? graphEntry.contentTypeId);
-        const key = graphEntry.tempId ?? `${graphEntry.contentTypeId}-${entryIndex}`;
-
-        return {
-          key,
-          entryIndex,
-          title: displayName,
-          contentType: displayName,
-          mappingCount: graphEntry.fieldMappings.length,
-        };
-      }),
+    () => buildOverviewEntries(payload.entryBlockGraph.entries, payload.contentTypes),
     [payload.contentTypes, payload.entryBlockGraph.entries]
   );
 
@@ -665,6 +520,7 @@ export const DocumentOutline = ({ payload, showChrome = true, onBack }: Document
   return (
     <Flex flexDirection="column" gap="spacingM" style={{ padding: tokens.spacingL }}>
       {showChrome && (
+        // TODO: update this overview to use the overview section component
         <Card padding="default" style={{ border: `1px solid ${tokens.gray300}` }}>
           <Flex flexDirection="column" gap="spacingS">
             <Box>
